backend/requirements.txt
# Framework Web e ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Redis e WebSockets
redis==5.0.1
websockets==12.0

# Validação de dados
pydantic==2.5.0
pydantic-settings==2.1.0

# Segurança e Autenticação
python-multipart==0.0.6
PyJWT==2.8.0 
passlib[bcrypt]==1.7.4

# Variáveis de ambiente
python-dotenv==1.0.0

# LLM Providers (conforme PDF)
openai==1.3.0
google-generativeai==0.3.0
anthropic==0.7.4

# HTTP Client para LLMs
httpx==0.25.2

# Testes (conforme PDF - pytest obrigatório)
pytest==7.4.3
pytest-asyncio==0.21.1

# Linting e Code Quality (opcional mas recomendado)
ruff==0.1.6
mypy==1.7.1

# Monitoramento e Logs (bônus do PDF)
prometheus-client==0.17.1

# Utilitários
python-dateutil==2.8.2
pyyaml==6.0.1